+++ tests/connect_spec.lua
@@ -6,11 +6,12 @@ return {
 	run_test_async = function()
 		test_utils.ui_select_fake(1)
 		mssql.connect()
+		local client = vim.lsp.get_clients({ name = "mssql_ls", bufnr = 0 })[1]
 
 		-- The connect event is sent, then the intelliSenseReady event.
 		-- Wait for the intelliSenseReady event as this means the connection was successful and
 		-- We can progress to the next test (getting completion items)
-		local result, err = test_utils.wait_for_handler("textDocument/intelliSenseReady", 30000)
+		local result, err = test_utils.wait_for_handler(0, client, "textDocument/intelliSenseReady", 30000)
 		if err then
 			error(err.message)
 		end
